requirements.txt
addict==2.4.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
mmcv==1.3.10
mmsegmentation==0.15.0
numpy==1.24.4
opencv-python==4.11.0.86
packaging==25.0
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.11.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
timm==0.4.12
tomli==2.2.1
torch==1.9.0
torchvision==0.10.0
typing_extensions==4.13.2
wcwidth==0.2.13
yapf==0.43.0
zipp==3.20.2
